pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-ai-toolkit"
version = "1.0.0"
description = "A comprehensive toolkit for building agentic AI systems"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "agents", "llm", "langchain", "langgraph", "agentic"]

dependencies = [
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "langgraph>=0.2.0",
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
google = [
    "google-adk>=0.1.0",
    "google-genai>=0.1.0",
    "litellm>=1.0.0",
]
observability = [
    "langsmith>=0.1.0",
]
a2a = [
    "a2a-sdk>=0.3.0",
    "uvicorn>=0.24.0",
    "aiohttp>=3.9.0",
]
paper = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "numpy>=1.24.0",
]
all = [
    "agentic-ai-toolkit[dev,google,observability,a2a,paper]",
]

[project.urls]
Homepage = "https://github.com/example/agentic-ai-toolkit"
Documentation = "https://github.com/example/agentic-ai-toolkit/docs"
Repository = "https://github.com/example/agentic-ai-toolkit"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4"]
ignore = ["D100", "D104", "D107"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
